pyproject.toml
[tool.poetry]
authors = [
    "Alexey Tsivunin <mailto:mailto:35929293+birthdaysgift@users.noreply.github.com>",
]
description = "BFF for WLSS frontend."
name = "wlss-bff"
version = "0.1.0"


[tool.poetry.dependencies]
python = "~3.11"


[tool.poetry.group.app]
optional = true

[tool.poetry.group.app.dependencies]
api = {git = "https://github.com/week-password/wlss-backend.git", rev = "f722f39af91b39cebc433b3419c0211d9555d8e9"}
fastapi = {extras = ["all"], version = "0.109.0" }  # we need "all" at least for uvicorn
pydantic = "2.5.3"
pyjwt = "2.8.0"
varname = "0.12.0"
wlss = {git = "https://github.com/week-password/wlss-backend-lib.git", rev = "e2c11e344d096579e2582c45fb08007195ebcc8d"}


[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
flake8 = "7.0.0"
flake8-spellcheck = "0.28.0"
mypy = "1.2.0"
ruff = "0.2.2"


[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
mbtest = "2.12.0"
pytest = "8.0.1"
pytest-cov = "4.1.0"
pytest-env = "0.8.1"  # needed to set up default $WLSS_ENV for tests execution


[tool.poetry.group.deploy]
optional = true


[tool.poetry.group.deploy.dependencies]
gunicorn = "20.1.0"


[build-system]

requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
